Decouple M-mode and supervisor/user code in memory
Previously, the supervisor and user code for the tests was expected to
follow the M-mode code at LRAM address (0x7f80_0000).
Decoupled these regions so the infrastructure M-mode code is at the LRAM
address and the supervisor/user code can be anywhere.

Updated all the tests to place the supervisor/user code at the
DRAM address (0x8000_0000).