Fix test003 to not change ra and use a0 instead.
We are supposed to return from the registered trap
handler override and the main stvec_trap_handler is
supposed to deal with the return logic.

Here we were setting the ra to point to the test
code and we were doing a simple call return which
is wrong way of returning.

Signed-off-by: Rajnesh Kanwal <rkanwal@rivosinc.com>