Add Qemu support to end tests
QEMU mode can be enabled on the command line at setup time:
-Ddiag_attribute_overrides=in_qemu_mode=True

Qemu AP machine has a separate MMIO region (sifive test device) that
can be written to indicate pass/fail for the tests. Hook that up in jumpstart
so that we can get pass/fail status while running in fs-sim.

The default Qemu runs on our G-die configuration (4 harts). Thus, all
the jumpstart tests will not pass as they are not configured for 4 harts.

Only test13 and test15 pass with fs-sim as they are configured to run with
4 harts. No extra tests have been added as the success case verified with
those tests. Every other test will fail as expected.

Signed-off-by: Atish Patra <atishp@rivosinc.com>