Put inactive cores in wfi state
Any core that comes up but not configured in active hart mask result in
test failure. But we can just park them in wfi instead of failing.
If a diag is configured for 4 harts, we should make sure that all the
harts passes the diag. However, if it has more harts, the diag should just
ignore those harts by parking them in wfi.

Signed-off-by: Atish Patra <atishp@rivosinc.com>