Added checkTC enable/disable magic instructions
These magic instructions tell checkTC to disable/enable comparison
between the spike and RTL traces. We want to disable instruction by
instruction comparison for sections of the diag that may not run the
same across Spike and RTL - for example in sync routines where the harts
true share data locations.