Added support for nested exceptions.
Allocate space for MAX_NUM_CONTEXT_SAVES nested exceptions per mmode and
smode. We will push and pop the context as we enter and exit the
mtvec/stvec trap handlers.
The run_function_in_*mode() functions will also save the context of the
mode they are coming from in this region before jumping to the function
in the lower level to be run.

Added test040 to test smode function run during mmode exception.
Updated a few existing tests to test nested exceptions.

Signed-off-by: Jerin Joy <joy@rivosinc.com>