Minor cleanup of the run_function_in_smode/umode() code
Removed the need to save/restore a0/a1 passed to
run_functions_in_umode().

Updated run_functions_in_*mode() return path to patch the return value
in a0 into the context that was saved before the function was called.
This makes the code cleaner for the upcoming nested exceptions change.

Signed-off-by: Jerin Joy <joy@rivosinc.com>