uart: Add pr_debug macro to only print in debug mode.
When running on rtl, debug prints seem to consume a lot
of cycles. After the initial test development the prints
are often no longer useful. This macro will allow
to keep the prints in for any future debugging and won't
add a penalty when running diags on rtl in release mode.

Signed-off-by: Rajnesh Kanwal <rkanwal@rivosinc.com>