common/uart: printk should not crash test when UART disabled
When the `disable_uart` attribute is set, a test should not
initialise or otherwise use the UART. However, since this attribute
might be set arbitrarily, it should not prevent a test from touching
printk().  (For example, default to no debug but printk()s are present
and can be enabled for debugging.)

This patch stops printk() from bombing out if the UART is not
initialised.  (This common code cannot currently check `disable_uart`,
which isn't a commoon attribute; future work could move this and
assert/error on a !initialised && !disabled condition.)