Fixed bare translation mode support.
No page table mappings are created for the stage if the mode is set to
bare.
Added checks to make sure that bare mappings have no_pte_allocation set
as well as no source address.
Page table mappings are also not generated for the jumpstart memory
sections for a particular stage in bare mode.

Added test051 to test for bare mode.

Signed-off-by: Jerin Joy <joy@rivosinc.com>