Added diag_custom_defines meson option
Allows diag specific defines to be overridden at meson setup time.

The diag source needs to provide default values for the defines:

build_diag.py takes the custom defines for the diag as a list:

./scripts/build_diag.py \
  --diag_custom_defines MAGIC_NUMBER=0xcafe ANOTHER_MAGIC_NUMBER=0xbeef

This is passed to meson setup as:
meson setup builddir \
  -Ddiag_custom_defines="-DMAGIC_NUMBER=0xcafe
  -ANOTHER_MAGIC_NUMBER=0xbeef"

build_diag.py passes these to DiagBuildTarget() using the
args.override_meson_options. We can do the same for the custom_rcode and
the override_diag_attributes.

Signed-off-by: Jerin Joy <joy@rivosinc.com>