DiagBuildTarget.add_build_asset() will copy by default
If the ELF is moved into the --diag_build_dir after the meson compile
step the meson test step will have to relink the ELF.
The relinking is happening for spike runs but not for qemu runs because
the diag ELF dependency is correctly handled by the meson test call for
spike.

The right thing to do is to copy the ELF and not move it after meson
compile so that meson test can just use that ELF.

The ELF dependency issue for the meson test qemu call will be fixed in
the next commit.

Signed-off-by: Jerin Joy <joy@rivosinc.com>