Updated meson test to add a dependency for diag_exe
For the spike target the diag_exe dependency was correctly enforced
because we're using it directly.
For oswis and qemu the dependency wasn't enforced because we're using
diag_exe.full_path().

This was causing issues with build_diag.py flow where we would
(incorrectly) move the ELF to the diag_build_dir after the compile step
instead of copying it. build_diag.py is fixed to copy instead of move
but it's better to enforce this dependency.

Signed-off-by: Jerin Joy <joy@rivosinc.com>