Only build in sources of modes that are enabled.
Added ifdefs to gate references in a higher mode to a lower mode that
isn't enabled in the build.

Renamed smode data area to privileged data area.
This is shared by both mmode and smode. We're going to allow builds
without smode regions but we still need this data area for mmode.
Include this privileged data area along with the mmode area.

Signed-off-by: Jerin Joy <joy@rivosinc.com>