Removed test043
This was used to test the code in the mmode section being larger than
the number of pages allocated for it.

This is not needed anymore because generate_diag_sources.py will add a
guard linker section after every linker section that isn't followed by
another section. When a section is overrun and runs into the guard
section the linker will fail the build.

This test required num_pages_for_jumpstart_mmode_text to have more pages
than were actually required just to so that it could build. We can now
reduce num_pages_for_jumpstart_mmode_text.

Signed-off-by: Jerin Joy <joy@rivosinc.com>