Add load_reserved and store conditional macros.
These macros are helpful in multicore senarios
where all cores share same data and need to operate
on it.

This can be done with spin lock as well but that
requires extra lock variable and cores spin around
it where as with these we can avoid most of the
extra logic. See the next commit for an example.

Signed-off-by: Rajnesh Kanwal <rkanwal@rivosinc.com>