mmode: Handle inactive CPUs as early as possible
Context:
We eventually want only the active cores accessing per core data
structures. This will allow us to reduce the data structures to only
cover the active cores and not max_num_cores_supported.

The quicker we dispatch the inactive cores the better. This is not
complete because when running in batch mode the inactive cores require a
stack to be set up to exit.

Signed-off-by: Jerin Joy <joy@rivosinc.com>